feat: auto-focus when single suggestion
Adds a feature that auto-highlights the suggestion when only one
suggestion is returned.
This allows the user to more naturally select with the keyboard - rather
than having to press the down/up arrow, they can press enter or tab
(assuming selectOnTabOut is true)

Added new autoFocusSingleSelection property with default true to allow
disabling of this behavior if desired.