fix: arrow up/down highlighting moves off list
When using keyboard arrow up/down to highlight items where there are
more suggestions than can be displayed by the maxResults parameter,
the highlight disappears but is still selecting an item, allowing
selection of a non-visible suggestion.

This limits keyboard highlighting to the lesser of the number of
suggestions or the value of maxResults.